MoneyBox Exercise
This is the solution of the Moneybox Android Technical Challenge. The user is able to login using username and password and see the account information about ISA and GIA accounts. Also the user is able to select specific account and make 10 pounds payment. The application is connected to server to retrieve the data so you need internet connection to use this app.